refactor(team): extract stable active members and pending replies store from TeamDetailView
Move two self-contained concerns out of TeamDetailView.tsx into sibling
modules. Both are pure code motion: the moved bodies are byte-identical
to their previous form (verified by diffing against origin/main), the
only edits are the added `export` keywords, the new modules' own import
headers, and three import lines in TeamDetailView.tsx.

useStableActiveMembers.ts holds `useStableActiveMembers` plus its two
module-private comparators `areResolvedMembersEqual` and
`areMemberMcpPoliciesEqual`, which had no other caller.

teamPendingRepliesStore.ts holds the `PendingRepliesUpdater` type,
`setPendingRepliesForTeam` and `useTeamPendingReplies`; the two
module-level caches, `getPendingRepliesSnapshot` and
`subscribePendingReplies` stay module-private.

One dev-only nuance worth naming: the two module-level `Map`s currently
live in the component module, so React Fast Refresh discards them on
every edit to TeamDetailView.tsx during development. After the move they
survive that HMR boundary. Production behaviour is identical.

TeamDetailView.tsx: 3832 -> 3697 lines (frozen cap 3839, unchanged).
No test file was modified; the existing suites pass as-is, which is the
motion proof.
